requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
executing==2.0.1
gunicorn==21.2.0
importlib-metadata==7.0.1
ipython==8.19.0
jedi==0.19.1
joblib==1.3.2
matplotlib-inline==0.1.6
numpy==1.26.3
packaging==23.1
parso==0.8.3
prompt-toolkit==3.0.43
psycopg2==2.9.6
pure-eval==0.2.2
Pygments==2.17.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
threadpoolctl==3.2.0
traitlets==5.14.1
tzdata==2023.3
watermark==2.4.3
wcwidth==0.2.12
whitenoise==6.5.0
zipp==3.17.0
